% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite.R
\name{addsyms}
\alias{addsyms}
\title{add gene symbols to a GRanges that uses ensembl identifiers}
\usage{
addsyms(x, EnsDb = EnsDb.Hsapiens.v75::EnsDb.Hsapiens.v75)
}
\arguments{
\item{x}{a GRanges assumed to have mcols element `gene_id` with ensembl gene identifiers}

\item{EnsDb}{an instance of EnsDb}
}
\value{
a GRanges with mcols column 'symbol'
}
\description{
add gene symbols to a GRanges that uses ensembl identifiers
}
\note{
At this time no checking of reference genome consistency is performed.
}
\examples{

}
