% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basicColData}
\alias{basicColData}
\title{basicColData: metadata about a small collection of files for demonstrating TxRegInfra}
\format{DataFrame from S4Vectors}
\usage{
basicColData
}
\description{
basicColData: metadata about a small collection of files for demonstrating TxRegInfra
}
\examples{
data(basicColData)
head(basicColData)
}
\keyword{datasets}
