% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite.R
\name{basicFormatter}
\alias{basicFormatter}
\title{operate on a character vector to derive a DataFrame, splitting on a tokenand retrieving first and last split fragments as 'base' and 'type' fields}
\usage{
basicFormatter(x, spltok = "_")
}
\arguments{
\item{x}{character vector}

\item{spltok}{token to use in strsplit}
}
\value{
a DataFrame instance
}
\description{
operate on a character vector to derive a DataFrame, splitting on a tokenand retrieving first and last split fragments as 'base' and 'type' fields
}
\examples{
some = c('Adipose_Subcutaneous_allpairs_v7_eQTL',
  'CD14_DS17215_hg19_FP',
  'CD19_DS17186_hg19_FP',
  'ENCFF001WGV_hg19_HS',
  'ENCFF994OCD_hg19_HS')
basicFormatter(some)
}
