% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_eQTL_granges}
\alias{demo_eQTL_granges}
\title{a GRanges instance with eQTL metadata returned by `sbov`}
\format{a GRanges instance}
\usage{
demo_eQTL_granges
}
\description{
a GRanges instance with eQTL metadata returned by `sbov`
}
\examples{
names(mcols(demo_eQTL_granges))
}
\keyword{datasets}
