% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_fimo_granges}
\alias{demo_fimo_granges}
\title{a list of GRanges instances with TF FIMO scores returned by `fimo_granges`}
\format{a list of GRanges instances}
\usage{
demo_fimo_granges
}
\description{
a list of GRanges instances with TF FIMO scores returned by `fimo_granges`
}
\examples{
names(mcols(demo_fimo_granges))
}
\keyword{datasets}
