% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dgf_meta}
\alias{dgf_meta}
\title{dgf_meta: metadata about a small collection of bed files for demonstrating TxRegInfra}
\format{data.frame}
\usage{
dgf_meta
}
\description{
dgf_meta: metadata about a small collection of bed files for demonstrating TxRegInfra
}
\examples{
data(dgf_meta)
head(dgf_meta)
}
\keyword{datasets}
