% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dnmeta}
\alias{dnmeta}
\title{metadata about DNaseI hotspots from ENCODE}
\format{data.frame}
\usage{
dnmeta
}
\description{
metadata about DNaseI hotspots from ENCODE
}
\examples{
data(dnmeta)
head(dnmeta[,1:10])
}
\keyword{datasets}
