% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{dnmetaApp}
\alias{dnmetaApp}
\title{support search of hotspot/peak data from ENCODE}
\usage{
dnmetaApp()
}
\value{
data.frame of metadat about selections
}
\description{
support search of hotspot/peak data from ENCODE
}
\examples{
if (interactive()) {
    oask = options()$example.ask
    if (askYesNo("start app?"))
           require("shiny")
           dnmetaApp()
    options(example.ask=oask)
    }
}
