% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite.R
\name{getDocumentFields}
\alias{getDocumentFields}
\title{determine the fields present in a txregnet document}
\usage{
getDocumentFields(rme, docTypeName = "type")
}
\arguments{
\item{rme}{instance of RaggedMongoExperiment}

\item{docTypeName}{character(1) telling the name of the column of colData(rme) that supplies information on document type}
}
\value{
a character vector
}
\description{
determine the fields present in a txregnet document
}
\examples{
getDocumentFields
}
