% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite.R
\name{getFieldNames}
\alias{getFieldNames}
\title{get names of fields in a collection in remote txregnet}
\usage{
getFieldNames(collection, check = TRUE, url = URL_txregInAWS(),
  db = "txregnet", limitn = 1)
}
\arguments{
\item{collection}{character(1) name of collection}

\item{check}{logical(1) if TRUE will verify that \code{coll} is present}

\item{url}{character(1) mongodb url}

\item{db}{character(1) mongodb db name}

\item{limitn}{numeric(1) number of records to probe to get field names}
}
\value{
a vector of strings
}
\description{
get names of fields in a collection in remote txregnet
}
\examples{
getFieldNames("CD34_DS12274_hg19_FP")
}
