% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsFiles}
\alias{hsFiles}
\title{hsFiles: metadata about a small collection of bed files for demonstrating TxRegInfra}
\format{DataFrame}
\usage{
hsFiles
}
\description{
hsFiles: metadata about a small collection of bed files for demonstrating TxRegInfra
}
\examples{
data(hsFiles)
hsFiles[1:4,1:4]
}
\keyword{datasets}
