% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBedToMongo.R
\name{importBedToMongo}
\alias{importBedToMongo}
\title{arrange import to mongo using mongoimport, setting up type and fields appropriately}
\usage{
importBedToMongo(path, collectionName, bedType = "narrowPeak",
  dbname = "db", importCmd = "mongoimport", host = "127.0.0.1")
}
\arguments{
\item{path}{path to bed file (not compressed)}

\item{collectionName}{name to use in mongodb}

\item{bedType}{one of 'narrowPeak', 'broadPeak', contact developers for other types if desired}

\item{dbname}{mongodb database name, used directly with system("mongoimort ...")}

\item{importCmd}{how to invoke 'mongoimport', default is to assume it can be found in PATH}

\item{host}{host identifier for mongoimport, defaults to 127.0.0.1}
}
\description{
arrange import to mongo using mongoimport, setting up type and fields appropriately
}
\examples{
f1 = dir(system.file("bedfiles", package="TxRegInfra"), full=TRUE)[1]
importBedToMongo(f1, "vjc1", db="txregnet")
}
