% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFIMO.R
\name{importFIMO}
\alias{importFIMO}
\title{utility to read FIMO outputs (after sorting and tabix indexing applied as in contents of biocfound-tfbs S3 bucket)}
\usage{
importFIMO(tf, param, genomeTag = "hg19")
}
\arguments{
\item{tf}{character(1) file id}

\item{param}{a GRanges delimiting the extract}

\item{param}{a GRanges delimiting the extract}
}
\description{
utility to read FIMO outputs (after sorting and tabix indexing applied as in contents of biocfound-tfbs S3 bucket)
}
\examples{
requireNamespace("GenomicRanges")
fn = URL_s3_tf(tag="M5946")
importFIMO(fn, param=GenomicRanges::GRanges("chr1", IRanges(1,15000)))
}
