% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragged.R
\name{localMongolite}
\alias{localMongolite}
\title{create mongolite connection to local mongodb server}
\usage{
localMongolite(coll = "test", db = "test", url = "mongodb://127.0.0.1")
}
\arguments{
\item{coll}{character(1) collection name}

\item{db}{character(1) database name}

\item{url}{character(1) mongodb URL minus port}
}
\description{
create mongolite connection to local mongodb server
}
\examples{
if (verifyRunningMongodb()) print(localMongolite())
}
