% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite.R
\name{makeAggregator}
\alias{makeAggregator}
\title{generate JSON to aggregate (counting records, and, by default, averaging
a given variable) within a collection}
\usage{
makeAggregator(by = "chrom", vbl = "chromStart", opname = "average",
  op = "$avg")
}
\description{
generate JSON to aggregate (counting records, and, by default, averaging
a given variable) within a collection
}
\note{
This produces json that can be used as an argument to m$aggregate() for m a mongolite::mongo instance
}
\examples{
makeAggregator()
if (interactive()) {
   remURL = URL_txregInAWS()
   colls = listAllCollections( url=remURL, db = "txregnet")
   m1 = mongo(url = remURL, db = "txregnet", 
     collection="CD14_DS17215_hg19_FP")
# find minimum value of statistic 'stat' per chromosome
   newagg = makeAggregator( by="chr",
     vbl="stat", op="$min", opname="min")
   tab = m1$aggregate( newagg )
   head(tab)
   }
}
