% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite.R
\name{makeColData}
\alias{makeColData}
\title{generate a colData component corresponding to a mongodb}
\usage{
makeColData(url = URL_txregInAWS(), db = "txregnet",
  formatter = basicFormatter)
}
\arguments{
\item{url}{character(1) url for mongodb}

\item{db}{character(1) database name}

\item{formatter}{a function that takes in a character vector and
returns a DataFrame with number of rows equal to the length of input}
}
\value{
a DataFrame instance
}
\description{
generate a colData component corresponding to a mongodb
}
\examples{
if (verifyHasMongoCmd()) makeColData()
}
