% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite.R
\name{makeGRConverterList}
\alias{makeGRConverterList}
\title{generate a list of GRanges to JSON for queries to mongo}
\usage{
makeGRConverterList(rme, map = basicCfieldsMap(), docTypeName = "type")
}
\arguments{
\item{rme}{RaggedMongoExperiment instance}

\item{map}{list of lists of named character vectors}

\item{docTypeName}{character(1) that identifies sample 'type'}
}
\value{
a list of JSON documents
}
\description{
generate a list of GRanges to JSON for queries to mongo
}
