% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite.R
\name{mongoliteCon}
\alias{mongoliteCon}
\title{constructor for S4 representation of mongolite dbconnection}
\usage{
mongoliteCon(url, db, collection = "test")
}
\arguments{
\item{url}{character(1) mongodb reference}

\item{db}{character(1) name of database in use}

\item{collection}{character(1) collection for queries}
}
\description{
constructor for S4 representation of mongolite dbconnection
}
\examples{
if (interactive()) {
   tcon = mongoliteCon(url=URL_txregInAWS(), db="txregnet", collection="CD14_DS17215_hg19_FP")
   tcon
}
}
