% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oldbasicColData}
\alias{oldbasicColData}
\title{oldbasicColData: metadata about a small collection of files for demonstrating TxRegInfra}
\format{DataFrame from S4Vectors}
\usage{
oldbasicColData
}
\description{
oldbasicColData: metadata about a small collection of files for demonstrating TxRegInfra
}
\examples{
data(oldbasicColData)
head(oldbasicColData)
}
\keyword{datasets}
