% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragged.R
\name{RaggedMongoExpt}
\alias{RaggedMongoExpt}
\title{manage ragged interval data in mongodb collections using a RaggedExperiment}
\usage{
RaggedMongoExpt(con, colData = DataFrame())
}
\arguments{
\item{con}{should be instance of mongoliteCon or RMongo}

\item{colData}{S4Vectors DataFrame instance}
}
\description{
manage ragged interval data in mongodb collections using a RaggedExperiment
}
\examples{
\dontrun{
if (verifyRunningMongodb()) {
  okdf = DataFrame(hsFiles_subset) # creates hsFiles
  rownames(okdf) = hsFiles[,1]
  rme1 = RaggedMongoExpt(con, colData=okdf)
  rme1
  }
}
}
