% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbov_output_eQTL}
\alias{sbov_output_eQTL}
\title{a GRanges instance with eQTL metadata returned by `sbov`}
\format{a GRanges instance}
\usage{
sbov_output_eQTL
}
\description{
a GRanges instance with eQTL metadata returned by `sbov`
}
\examples{
head(sbov_output_eQTL)
}
\keyword{datasets}
