% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbov_output_FP}
\alias{sbov_output_FP}
\title{a GRanges instance with digital genomic footprint metadata returned by `sbov`}
\format{a GRanges instance}
\usage{
sbov_output_FP
}
\description{
a GRanges instance with digital genomic footprint metadata returned by `sbov`
}
\examples{
head(sbov_output_FP)
}
\keyword{datasets}
