% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbov_output_HS}
\alias{sbov_output_HS}
\title{a GRanges instance with DnaseI hotspot metadata returned by `sbov`}
\format{a GRanges instance}
\usage{
sbov_output_HS
}
\description{
a GRanges instance with DnaseI hotspot metadata returned by `sbov`
}
\examples{
head(sbov_output_HS)
}
\keyword{datasets}
