% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{sbov_to_graphNEL}
\alias{sbov_to_graphNEL}
\title{convert sbov() output to annotated graphNEL}
\usage{
sbov_to_graphNEL(sbovout, ...)
}
\arguments{
\item{sbovout}{output of sbov() when applied to a TxRegInfra eQTL resource from mongodb}

\item{\dots}{not used}
}
\value{
instance of graphNEL from Bioc graph package
}
\description{
convert sbov() output to annotated graphNEL
}
\note{
can convert to igraph via igraph::igraph.from.graphNEL
}
\examples{
sbov_to_graphNEL(demo_eQTL_granges)
}
