% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\docType{methods}
\name{subsetByOverlaps,RaggedMongoExpt,GRanges-method}
\alias{subsetByOverlaps,RaggedMongoExpt,GRanges-method}
\title{extract records from mongo store based on GRanges query}
\usage{
\S4method{subsetByOverlaps}{RaggedMongoExpt,GRanges}(x, ranges, collectionType,
  maxgap = -1L, minoverlap = 0L, type = c("any", "start", "end", "within",
  "equal"), invert = FALSE, ...)
}
\arguments{
\item{x}{instance of RaggedMongoExpt}

\item{ranges}{instance of GRanges}

\item{collectionType}{character(1) must be "bed" at this time}

\item{maxgap}{not used}

\item{minoverlap}{not used}

\item{type}{not used}

\item{invert}{not used}

\item{\dots}{not used}
}
\value{
instance of \code{\link[RaggedExperiment]{RaggedExperiment-class}}
}
\description{
extract records from mongo store based on GRanges query
}
\examples{
data(hsFiles)
okdf = DataFrame(hsFiles)
rownames(okdf) = hsFiles[,1]
rme1 = RaggedMongoExpt(dbName="txregnet", colData=okdf)
}
