% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txmodels.R
\name{txmodels}
\alias{txmodels}
\title{use Gviz to render transcript models via GeneRegionTrack, but keep lightweight through requireNamespace and suggestion for installation}
\usage{
txmodels(sym, gr, edb = "EnsDb.Hsapiens.v75", plot.it = FALSE,
  radius = 0, ...)
}
\arguments{
\item{sym}{a gene symbol to be looked up in biovizBase::genesymbol table}

\item{gr}{a GRanges instance, anticipated to be length 1}

\item{edb}{a character(1) name of an EnsDb annotation package}

\item{plot.it}{a logical(1) specifying whether Gviz::plotTracks should be run}

\item{radius}{a numeric(1) specifying number to add to IRanges instance used to subset gene models from ensembldb::exonsBy output}

\item{\dots}{passed to Gviz::GeneRegionTrack}
}
\value{
an instance of Gviz::GeneRegionTrack, invisibly returned
}
\description{
use Gviz to render transcript models via GeneRegionTrack, but keep lightweight through requireNamespace and suggestion for installation
}
\examples{
t0 = txmodels('ORMDL3', plot.it=TRUE, name='ORMDL3')
t1 = txmodels('ORMDL3', plot.it=FALSE, name='meta', collapseTranscripts='meta')
requireNamespace('Gviz')
Gviz::plotTracks(list(Gviz::GenomeAxisTrack(), t0, t1), showId=TRUE)
}
