% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWS.R
\name{txregRemote}
\alias{txregRemote}
\title{return mongolite connection for working mongo server}
\usage{
txregRemote(...)
}
\arguments{
\item{\dots}{passed to mongo()}
}
\description{
return mongolite connection for working mongo server
}
\examples{
if (interactive()) {
txregRemote()
}
}
