% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFIMO.R
\name{URL_s3_tf}
\alias{URL_s3_tf}
\title{utility to generate link to biocfound bucket for FIMO TFBS scores}
\usage{
URL_s3_tf(tag = "M3433")
}
\arguments{
\item{tag}{character(1) token identifying TF}
}
\description{
utility to generate link to biocfound bucket for FIMO TFBS scores
}
\examples{
URL_s3_tf
}
