% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWS.R
\name{URL_txregInAWS}
\alias{URL_txregInAWS}
\title{return mongodb URL for working mongo server}
\usage{
URL_txregInAWS()
}
\value{
character(1) URL for a hosted resource
}
\description{
return mongodb URL for working mongo server
}
\examples{
URL_txregInAWS
}
