% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite.R
\name{verifyHasMongoCmd}
\alias{verifyHasMongoCmd}
\title{check for existence of 'mongo' command, for db.getCollectionNames etc.}
\usage{
verifyHasMongoCmd()
}
\value{
logical(1)
}
\description{
check for existence of 'mongo' command, for db.getCollectionNames etc.
}
\examples{
verifyHasMongoCmd()
}
