% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite.R
\name{verifyRunningMongodb}
\alias{verifyRunningMongodb}
\title{check for accessible local mongodb}
\usage{
verifyRunningMongodb(url = "mongodb://127.0.0.1")
}
\arguments{
\item{url}{character(1) defining mongodb server}
}
\value{
logical(1)
}
\description{
check for accessible local mongodb
}
\examples{
if (interactive()) verifyRunningMongodb()
}
