% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite.R, R/queryBed.R
\name{grConverter}
\alias{grConverter}
\alias{grConverter}
\title{convert a GRanges to a JSON query for mongodb}
\usage{
grConverter(queryGRange, cfields = c(chrom = "chrom", start = "chromStart",
  end = "chromEnd"))

grConverter(queryGRange, cfields = c(chrom = "chrom", start = "chromStart",
  end = "chromEnd"))
}
\arguments{
\item{queryGRange}{a \code{\link[GenomicRanges]{GRanges-class}} instance of length 1}

\item{cfields}{a named character(3) vector with names 'chrom', 'start',
'end'; the element values will be used to name document fields in the query}

\item{queryGRange}{a length 1 GRanges instance specifying interval within which records are to be retrieved}

\item{cfields}{a vector with named elements 'chrom', 'start', 'end' indicating how the JSON components should be named to query fields in the target mongodb document}
}
\description{
convert a GRanges to a JSON query for mongodb

convert a GRanges instance to JSON suitable for RMongo::dbGetQuery
importFrom GenomicRanges start end
}
\examples{
gr = GRanges("chr1", cfields=c(chrom="chr", start="start", end="end")
require(GenomicRanges)
grConverter(GRanges("chr1", IRanges(1, 5000)))
}
