% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBedToMongo.R
\name{grConverter}
\alias{grConverter}
\title{convert a GRanges instance to JSON suitable for RMongo::dbGetQuery
import GenomeInfoDb
importFrom GenomicRanges start end}
\usage{
grConverter(queryGRange, cfields = c(chrom = "chrom", start = "chromStart",
  end = "chromEnd"))
}
\arguments{
\item{queryGRange}{a length 1 GRanges instance specifying interval within which records are to be retrieved}

\item{cfields}{a vector with named elements 'chrom', 'start', 'end' indicating how the JSON components should be named to query fields in the target mongodb document}
}
\description{
convert a GRanges instance to JSON suitable for RMongo::dbGetQuery
import GenomeInfoDb
importFrom GenomicRanges start end
}
\examples{
require(GenomicRanges)
grConverter(GRanges("chr1", IRanges(1, 5000)))
}
