% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBedToMongo.R
\name{importBedToMongo}
\alias{importBedToMongo}
\title{arrange import to mongo using mongoimport, setting up type and fields appropriately}
\usage{
importBedToMongo(path, collectionName, bedType = "narrowPeak",
  dbname = "db", importCmd = "mongoimport", host = "127.0.0.1")
}
\arguments{
\item{path}{path to bed file (not compressed)}

\item{collectionName}{name to use in mongodb}

\item{bedType}{one of 'narrowPeak', 'broadPeak', 'chromHMM': contact developers for other types if desired}

\item{dbname}{mongodb database name, used directly with system("mongoimort ...")}

\item{importCmd}{how to invoke 'mongoimport', default is to assume it can be found in PATH}

\item{host}{host identifier for mongoimport, defaults to 127.0.0.1}
}
\value{
if error encountered, return the try-error content, otherwise TRUE
}
\description{
arrange import to mongo using mongoimport, setting up type and fields appropriately
}
\examples{
f1 = dir(system.file("bedfiles", package="TxRegInfra"), full=TRUE, patt="ENCFF971VCD")
f2 = dir(system.file("bedfiles", package="TxRegInfra"), full=TRUE, patt="E096_imp12")
if (verifyHasMongoCmd("mongoimport")) {
chk1 = importBedToMongo(f1, "vjc1", db="txregnet")
stopifnot(chk1)
chk2 = importBedToMongo(f2, "vjc2", db="txregnet", bedType="chromHMM")
stopifnot(chk2)
system('mongo txregnet --eval "db.vjc1.remove({})"') # cleanup
system('mongo txregnet --eval "db.vjc2.remove({})"') # cleanup
}
}
