% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite.R
\name{listAllCollections}
\alias{listAllCollections}
\title{list all collections in a database, using command-line interface}
\usage{
listAllCollections(url = "mongodb://127.0.0.1:27017", db = "test")
}
\arguments{
\item{url}{character(1) mongodb URL}

\item{db}{character(1) mongodb database name}
}
\description{
list all collections in a database, using command-line interface
}
\examples{
if (verifyRunningMongodb()) listAllCollections()
}
