% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{class}
\name{RaggedMongoExpt-class}
\alias{RaggedMongoExpt-class}
\alias{RaggedMongoExpt}
\title{define container for mongodb backed RaggedExperiment}
\usage{
RaggedMongoExpt(dbName, colData = DataFrame())
}
\arguments{
\item{dbName}{character(1) name used for db in mongodb}

\item{colData}{a \code{\link[S4Vectors]{DataFrame-class}} instance}
}
\description{
define container for mongodb backed RaggedExperiment

define constructor for mongodb backed RaggedExperiment
}
