% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R, R/ragged.R
\docType{class}
\name{RaggedMongoExpt-class}
\alias{RaggedMongoExpt-class}
\alias{RaggedMongoExpt}
\alias{RaggedMongoExpt}
\title{define container for mongodb backed RaggedExperiment}
\usage{
RaggedMongoExpt(con, colData = DataFrame())

RaggedMongoExpt(con, colData = DataFrame())
}
\arguments{
\item{con}{should be instance of mongoliteCon or RMongo}

\item{colData}{a \code{\link[S4Vectors]{DataFrame-class}} instance}

\item{dbName}{character(1) name used for db in mongodb}

\item{colData}{S4Vectors DataFrame instance}
}
\description{
define container for mongodb backed RaggedExperiment

define constructor for mongodb backed RaggedExperiment

manage ragged interval data in mongodb collections using a RaggedExperiment
}
\examples{
\dontrun{
if (verifyRunningMongodb()) {
  okdf = DataFrame(hsFiles_subset) # creates hsFiles
  rownames(okdf) = hsFiles[,1]
  rme1 = RaggedMongoExpt(con, colData=okdf)
  rme1
  }
}
}
