% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite.R
\name{RaggedMongoExpt}
\alias{RaggedMongoExpt}
\title{bind colData to a mongo-based ragged-experiment incubator}
\usage{
RaggedMongoExpt(con, colData)
}
\arguments{
\item{con}{a mongolite::mongo instance}

\item{colData}{a DataFrame instance}
}
\value{
instance of RaggedMongoExpt
}
\description{
bind colData to a mongo-based ragged-experiment incubator
}
