% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite.R
\name{sbov}
\alias{sbov}
\title{prototype of subsetter for mongo resource}
\usage{
sbov(rme, gr, map = basicCfieldsMap(), docTypeName = "type")
}
\arguments{
\item{rme}{RaggedMongoExpt instance}

\item{gr}{GRanges instance to subset by}

\item{map}{list with one element per document type telling what fields are chr, start, stop}

\item{docTypeName}{character(1) naming column of colData(rme) that has document type}
}
\value{
a RaggedExperiment instance
}
\description{
prototype of subsetter for mongo resource
}
\examples{
requireNamespace("mongolite")
if (verifyHasMongoCmd()) {  # for makeColData
 m1 = mongolite::mongo(url=URL_txregInAWS(), db="txregnet")
 cd = makeColData(url=URL_txregInAWS(), db="txregnet")
 rme1 = RaggedMongoExpt(m1, cd[which(cd$type=="FP"),][1:8,])
 BiocParallel::register(BiocParallel::SerialParam())
 ss = sbov(rme1, GRanges("chr1", IRanges(1e6, 1.5e6)))
}
}
