% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite.R
\name{txregCollections}
\alias{txregCollections}
\title{list collections in AWS mongo server for txregnet}
\usage{
txregCollections(ignore = 1:3, url = URL_txregInAWS(), db = "txregnet")
}
\arguments{
\item{ignore}{integer vector telling which lines of mongo db.getCollectionNames() result should be ignored}

\item{url}{a valid mongodb URL}

\item{db}{character(1) db name}
}
\description{
list collections in AWS mongo server for txregnet
}
\examples{
txregCollections()[1:5]
}
