% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite.R
\name{verifyHasMongoCmd}
\alias{verifyHasMongoCmd}
\title{check for existence of 'mongo' command, for db.getCollectionNames etc.}
\usage{
verifyHasMongoCmd(cmd = "mongo")
}
\arguments{
\item{cmd}{character(1) either 'mongo' or 'mongoimport'}
}
\value{
logical(1)
}
\description{
check for existence of 'mongo' command, for db.getCollectionNames etc.
}
\note{
we use mongoimport command to import tsv files; mongolite import 'method' not immediately useful for this
}
\examples{
if (interactive()) verifyHasMongoCmd()
}
